backend/pyproject.toml
[project]
name = "shadow-agent-map"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.29.0",
    "google-adk>=0.3.0",
    "google-genai>=1.0.0",
    "google-cloud-bigquery>=3.20.0",
    "google-cloud-run>=0.10.0",
    "google-cloud-aiplatform>=1.60.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.27.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
